Add axis support with unique_with_counts
This fix tries to address the issue raised in 15644 where
unique_with_counts does not take an axis and the input tensor
is only limited to 1-D.

This fix registers the op/kernel of UniqueWithCountsV2
so that the axis could be passed.

The axis support for tf.unique_with_counts in python
has not been enabled yet. Instead, this fix adds a python
wrapper to points to `gen_array_ops._unique_with_counts` and
hides `UniqueWithCounts` for now, so that
it is possible to switch to `gen_array_ops._unique_with_counts_v2`
once the API deprecation process is complete (3 weeks).

This fix is part of the effort to fix 15644.

Signed-off-by: Yong Tang <yong.tang.github@outlook.com>